Revert "Cineform: Windows warning fix"
This reverts commit 45d0bfb61583986b2b0efdd87530e9e39768c8f1.

+++ src/video_compress/cineform.cpp
@@ -340,7 +340,7 @@ static video_frame *get_copy(struct state_video_compress_cineform *s, video_fram
                 dst += static_cast<size_t>(dst_linesize) * frame->tiles[0].height - 1;
                 dst_sign = -1;
         }
-        for (unsigned i = 0; i < frame->tiles[0].height; ++i) {
+        for (long int i = 0; i < frame->tiles[0].height; ++i) {
                 s->dec(dst + dst_sign * i * dst_linesize, (unsigned char *) frame->tiles[0].data + i * src_linesize, dst_linesize, 16, 8, 0);
         }
